refactor: issue-311 フロントエンドテストコードの統合と最適化
テストコードの重複を削減し、保守性を向上させるため以下を実施:
- API系テスト: 重複テストを削除し、パラメータ化を活用（13%削減）
- Logger系テスト: 5ファイルから3ファイルに統合（17%削減）
- Hooks系テスト: スキップされたテストを削除（11%削減）
全体で484行（14.5%）のコード削減を達成
