--- frontend/src/lib/logger/__tests__/config.test.ts
@@ -1,164 +1,0 @@
-/**
- * 設定管理テスト（最適化版）
- *
- * 重要な設定検証のみに焦点を絞り、実装詳細への依存を削減
- */
-
-import { describe, expect, it } from "vitest";
-import {
-	getDefaultConfig,
-	mergeConfigs,
-	shouldLog,
-	validateConfig,
-} from "../config";
-import type { BrowserLoggerConfig, LogLevel } from "../types";
-
-describe("Logger Config", () => {
-	describe("getDefaultConfig", () => {
-		it("各環境で適切なデフォルト設定を返す", () => {
-			const devConfig = getDefaultConfig("development");
-			expect(devConfig.level).toBe("debug");
-			expect(devConfig.enableConsole).toBe(true);
-			expect(devConfig.bufferSize).toBe(10);
-
-			const prodConfig = getDefaultConfig("production");
-			expect(prodConfig.level).toBe("info");
-			expect(prodConfig.enableConsole).toBe(false);
-			expect(prodConfig.bufferSize).toBe(100);
-
-			const storybookConfig = getDefaultConfig("storybook");
-			expect(storybookConfig.level).toBe("warn");
-			expect(storybookConfig.enableConsole).toBe(true);
-		});
-
-		it("必須フィールドが全て含まれる", () => {
-			const config = getDefaultConfig("development");
-			const requiredFields = [
-				"level",
-				"bufferSize",
-				"flushInterval",
-				"maxRetries",
-				"enableConsole",
-			];
-
-			requiredFields.forEach((field) => {
-				expect(config).toHaveProperty(field);
-			});
-		});
-	});
-
-	describe("validateConfig", () => {
-		it("有効な設定を受け入れる", () => {
-			const validConfig = getDefaultConfig("production");
-			const result = validateConfig(validConfig);
-			expect(result.valid).toBe(true);
-			expect(result.errors).toHaveLength(0);
-		});
-
-		it("必須フィールドが欠けている場合の検証", () => {
-			// validateConfigは値の範囲のみを検証し、必須フィールドの存在はチェックしない
-			// 不完全なオブジェクトを渡した場合、undefinedの値に対して検証が実行される
-			const partialConfig = {
-				bufferSize: -1, // 無効な値
-			} as unknown as BrowserLoggerConfig;
-
-			const result = validateConfig(partialConfig);
-			expect(result.valid).toBe(false);
-			expect(result.errors).toContain("bufferSize must be between 1 and 1000");
-		});
-
-		it("無効な値でエラー", () => {
-			const baseConfig = getDefaultConfig("development");
-
-			// 無効な値のテスト - validateConfigはログレベルを検証しないので、数値範囲のテストに変更
-			const result = validateConfig({
-				...baseConfig,
-				bufferSize: 1001, // 最大値を超える
-			});
-			expect(result.valid).toBe(false);
-			expect(result.errors).toContain("bufferSize must be between 1 and 1000");
-
-			// 負の値
-			const negativeBufferResult = validateConfig({
-				...baseConfig,
-				bufferSize: -1,
-			});
-			expect(negativeBufferResult.valid).toBe(false);
-			expect(negativeBufferResult.errors.length).toBeGreaterThan(0);
-
-			// 無効なフラッシュインターバル
-			const negativeIntervalResult = validateConfig({
-				...baseConfig,
-				flushInterval: -1,
-			});
-			expect(negativeIntervalResult.valid).toBe(false);
-			expect(negativeIntervalResult.errors.length).toBeGreaterThan(0);
-		});
-	});
-
-	describe("shouldLog", () => {
-		const testCases: Array<[LogLevel, LogLevel, boolean]> = [
-			// [設定レベル, ログレベル, 期待値]
-			["debug", "debug", true],
-			["debug", "info", true],
-			["info", "debug", false],
-			["info", "info", true],
-			["warn", "info", false],
-			["warn", "warn", true],
-			["error", "warn", false],
-			["error", "error", true],
-		];
-
-		testCases.forEach(([configLevel, logLevel, expected]) => {
-			it(`設定${configLevel}でログ${logLevel}は${expected ? "出力" : "スキップ"}`, () => {
-				expect(shouldLog(configLevel, logLevel)).toBe(expected);
-			});
-		});
-	});
-
-	describe("mergeConfigs", () => {
-		it("デフォルト設定と部分的な設定をマージ", () => {
-			const base = getDefaultConfig("development");
-			const partial = {
-				level: "warn" as LogLevel,
-				bufferSize: 20,
-			};
-
-			const merged = mergeConfigs(base, partial);
-			expect(merged.level).toBe("warn");
-			expect(merged.bufferSize).toBe(20);
-			expect(merged.enableConsole).toBe(base.enableConsole); // 変更されていない
-		});
-
-		it("複数の設定を順番にマージ", () => {
-			const base = getDefaultConfig("development");
-			const override1 = {
-				level: "info" as LogLevel,
-				bufferSize: 20,
-			};
-			const override2 = {
-				level: "warn" as LogLevel,
-				enableConsole: false,
-			};
-
-			const merged = mergeConfigs(base, override1) as BrowserLoggerConfig;
-			const final = mergeConfigs(merged, override2) as BrowserLoggerConfig;
-			expect(final.level).toBe("warn"); // 最後の値が優先
-			expect(final.bufferSize).toBe(20); // override1から
-			expect(final.enableConsole).toBe(false); // override2から
-		});
-
-		it("深いオブジェクトのマージ", () => {
-			const base = getDefaultConfig("development");
-			const override = {
-				level: "error" as LogLevel,
-				bufferSize: 100,
-			};
-
-			const merged = mergeConfigs(base, override) as BrowserLoggerConfig;
-			expect(merged.level).toBe("error");
-			expect(merged.bufferSize).toBe(100);
-			expect(merged.flushInterval).toBe(base.flushInterval);
-		});
-	});
-});
